restructures app.js and auth.js(former login.js)
