+++ app.js
@@ -1,20 +1,22 @@
 require('dotenv').config();
 
-const bodyParser   = require('body-parser');
+const bodyParser = require('body-parser');
 const cookieParser = require('cookie-parser');
-const express      = require('express');
-const favicon      = require('serve-favicon');
-const hbs          = require('hbs');
-const mongoose     = require('mongoose');
-const logger       = require('morgan');
-const User         = require('./models/user');
-const path         = require('path');
-//const passport     = require('passport');
-const bcrypt       = require('bcryptjs');
-const SpotifyStrategy = require('passport-spotify').Strategy;
+const express = require('express');
+const favicon = require('serve-favicon');
+const hbs = require('hbs');
+const mongoose = require('mongoose');
+const logger = require('morgan');
+const User = require('./models/user');
+const path = require('path');
+const app = express();
+const bcrypt = require('bcryptjs');
+
 
 mongoose
-  .connect('mongodb://localhost/204U', {useNewUrlParser: true})
+  .connect('mongodb://localhost/204U', {
+    useNewUrlParser: true
+  })
   .then(x => {
     console.log(`Connected to Mongo! Database name: "${x.connections[0].name}"`)
   })
@@ -25,82 +27,34 @@ mongoose
 const app_name = require('./package.json').name;
 const debug = require('debug')(`${app_name}:${path.basename(__filename).split('.')[0]}`);
 
-const app = express();
-
-//Passport middleware configuration 
-const passport = require('passport');
-
-passport.serializeUser(  function(user, cb) { cb(null, user); });
-passport.deserializeUser(function(obj,  cb) { cb(null, obj);  });
-
-const client_id = 'a3676c8b791c49048c222a84f7fd770c';
-const client_secret = '54708907189e407e8c5f6eb2328f9374';
-
-/* GET personal Signup page */
-// router.get('/', (req, res, next) => {
-//   console.log("WE ARE AT LOGIN")
-//   res.render('login');
-// });
-
-// passport-spotify auth
-passport.use(
-  new SpotifyStrategy(
-    {
-      clientID: client_id,
-      clientSecret: client_secret,
-      callbackURL: 'http://127.0.0.1:3000/auth/spotify/callback'
-    },
-    function(accessToken, refreshToken, expires_in, profile, done) {
-      User.find({ spotifyId: profile.id }, function(err, user) {
-        if (user === null) {
-          User.create({ spotifyId: profile.id }).then((user) => {
-            return done(err, user); 
-          })
-        } else {
-          return done(err, user); 
-        }
-      });
-    }
-  )
-);
-app.use(passport.initialize());
-app.use(passport.session());
-
-var session = require ('express-session')
-app.use(session({
-  secret:"204u"
-}));
 // Middleware Setup
 app.use(logger('dev'));
 app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.urlencoded({
+  extended: false
+}));
 app.use(cookieParser());
 
 // Express View engine setup
-
 app.use(require('node-sass-middleware')({
-  src:  path.join(__dirname, 'public'),
+  src: path.join(__dirname, 'public'),
   dest: path.join(__dirname, 'public'),
   sourceMap: true
 }));
-      
 
 app.set('views', path.join(__dirname, 'views'));
 app.set('view engine', 'hbs');
 app.use(express.static(path.join(__dirname, 'public')));
 app.use(favicon(path.join(__dirname, 'public', 'images', 'favicon.ico')));
 
-
-
 // default value for title local
 app.locals.title = '20-4-U';
 
-// random comment
-
+// Route Handler
 const home = require('./routes/home');
 app.use('/', home);
 
-const login = require('./routes/login');
-app.use('/login', login);
+const auth = require('./routes/auth');
+app.use('/auth', auth);
 
 module.exports = app;
